+++ corehq/apps/users/urls.py
@@ -28,8 +28,6 @@ from .views import (
     test_httpdigest,
     undo_remove_web_user,
     verify_phone_number,
-    delete_domain_permission_mirror,
-    create_domain_permission_mirror,
     download_web_users,
     DownloadWebUsersStatusView,
     WebUserUploadJobPollView,
@@ -119,12 +117,6 @@ urlpatterns = [
         name=WebUserUploadJobPollView.urlname),
     url(r'^enterprise/$', EnterpriseUsersView.as_view(), name=EnterpriseUsersView.urlname),
     url(r'^enterprise/json/$', paginate_enterprise_users, name='paginate_enterprise_users'),
-    url(r'^enterprise/domain_permission_mirrors/$', DomainPermissionsMirrorView.as_view(),
-        name=DomainPermissionsMirrorView.urlname),
-    url(r'^enterprise/delete_domain_permission_mirror/(?P<mirror>[ \w-]+)/$', delete_domain_permission_mirror,
-        name='delete_domain_permission_mirror'),
-    url(r'^enterprise/create_domain_permission_mirror/$', create_domain_permission_mirror,
-        name='create_domain_permission_mirror'),
     url(r'^join/(?P<uuid>[ \w-]+)/$', accept_invitation, name='domain_accept_invitation'),
     url(r'^roles/$', ListRolesView.as_view(), name=ListRolesView.urlname),
     url(r'^roles/save/$', post_user_role, name='post_user_role'),
